wip: ctm function naming
