+++ tests/peps/test_environments.py
@@ -28,7 +28,7 @@ def test_finite_spinless_boundary_mps_ctmrg():
     fid, fc, fcdag = ops.I(), ops.c(), ops.cp()
 
     g_hop = fpeps.gates.gate_nn_hopping(t, dbeta / 2, fid, fc, fcdag)  # nn gate for 2D fermi sea
-    gates = fpeps.gates_homogeneous(geometry, gates_nn=g_hop)
+    gates = fpeps.gates.distribute(geometry, gates_nn=g_hop)
 
     psi = fpeps.product_peps(geometry, fid) # initialized at infinite temperature
     env = fpeps.EnvNTU(psi, which='NN+')
@@ -45,7 +45,8 @@ def test_finite_spinless_boundary_mps_ctmrg():
 
     env = fpeps.EnvCTM(psi)
     opts_svd_ctm = {'D_total': 30, 'tol': 1e-10}
-    for step in fpeps.ctmrg_(env, max_sweeps=50, iterator_step=1, opts_svd=opts_svd_ctm):
+    for _ in range(50):
+        env.update_(opts_svd=opts_svd_ctm)
         cdagc = env.measure_nn(fcdag, fc)
         energy = -2 * np.mean([*cdagc.values()])
 
@@ -57,9 +58,9 @@ def test_finite_spinless_boundary_mps_ctmrg():
     mpsenv = fpeps.EnvBoundaryMps(psi, opts_svd=opts_svd_ctm, setup='tlbr')
 
     for ny in range(psi.Ny):
-        vR0 = step.env.boundary_mps(n=ny, dirn='r')
+        vR0 = env.boundary_mps(n=ny, dirn='r')
         vR1 = mpsenv.boundary_mps(n=ny, dirn='r')
-        vL0 = step.env.boundary_mps(n=ny, dirn='l')
+        vL0 = env.boundary_mps(n=ny, dirn='l')
         vL1 = mpsenv.boundary_mps(n=ny, dirn='l')
 
         print(mps.vdot(vR0, vR1) / (vR0.norm() * vR1.norm()))  # problem with phase in peps?
@@ -68,9 +69,9 @@ def test_finite_spinless_boundary_mps_ctmrg():
         assert abs(abs(mps.vdot(vL0, vL1)) / (vL0.norm() * vL1.norm()) - 1) < 1e-7
 
     for nx in range(psi.Nx):
-        vT0 = step.env.boundary_mps(n=nx, dirn='t')
+        vT0 = env.boundary_mps(n=nx, dirn='t')
         vT1 = mpsenv.boundary_mps(n=nx, dirn='t')
-        vB0 = step.env.boundary_mps(n=nx, dirn='b')
+        vB0 = env.boundary_mps(n=nx, dirn='b')
         vB1 = mpsenv.boundary_mps(n=nx, dirn='b')
 
         print(mps.vdot(vT0, vT1) / (vT0.norm() * vT1.norm()))  # problem with phase in peps?
@@ -90,7 +91,7 @@ def test_spinless_infinite_approx():
     ops = yastn.operators.SpinlessFermions(sym='U1', backend=cfg.backend, default_device=cfg.default_device)
     fid, fc, fcdag = ops.I(), ops.c(), ops.cp()
     g_hop = fpeps.gates.gate_nn_hopping(t, dbeta / 2, fid, fc, fcdag)  # nn gate for 2D fermi sea
-    gates = fpeps.gates_homogeneous(geometry, gates_nn=g_hop)
+    gates = fpeps.gates.distribute(geometry, gates_nn=g_hop)
 
     psi = fpeps.product_peps(geometry, fid) # initialized at infinite temperature
     env = fpeps.EnvNTU(psi, which='NN+')
