+++ yastn/tn/fpeps/__init__.py
@@ -3,7 +3,7 @@ from ._peps import Peps
 from ._initialize import product_peps, load_from_dict
 from ._evolution import evolution_step_
 from ._doublePepsTensor import DoublePepsTensor
-from .gates import Gates, Gate_local, Gate_nn, gates_homogeneous
+from .gates import Gates, Gate_local, Gate_nn
 from ._ctmrg import ctmrg_
 from .envs._env_ctm import EnvCTM
 from .envs._env_ntu import EnvNTU
